% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{anyDuplicated.spp}
\alias{anyDuplicated.spp}
\title{Check for duplicated points in a spherical point pattern.}
\usage{
\method{anyDuplicated}{spp}(x, ...)
}
\arguments{
\item{x}{Spherical point patten of class \code{"spp"}.}

\item{...}{Additional arguments passed to
  \code{\link{anyDuplicated.data.frame}}.}
}
\value{
An integer or real vector of length one with value the 1-based index
  of the first duplicate if any, otherwise 0.
}
\description{
Check for duplicated points in a spherical point pattern.
}

