% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sphericaldomain.R
\name{area.sphericaldomain}
\alias{area.sphericaldomain}
\title{Area of spherical domain}
\usage{
\method{area}{sphericaldomain}(w)
}
\arguments{
\item{w}{Objcet of class \code{"sphericaldomain"}.}
}
\description{
Area of spherical domain
}
\examples{
dom <- sphere(2)
area(dom)
dom <- sphericalcap(c(0,0), pi/2, radius = 2)
area(dom)
}

