% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{as.data.frame.spp}
\alias{as.data.frame.spp}
\title{Extract coordinates of a spherical point pattern into a data.frame}
\usage{
\method{as.data.frame}{spp}(x, row.names = NULL, ..., coord_type = NULL)
}
\arguments{
\item{x}{Spherical point patten of class \code{"spp"}.}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{...}{Ignored.}

\item{coord_type}{String. Type of coordinates used for return coordinates. Defaults to coordinate type of \code{x}.
 See \code{\link{spp}} for supported coordinate types.}
}
\value{
\code{data.frame} with coordinates.
}
\description{
Extract coordinates of a spherical point pattern into a data.frame
}

