% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericaldomain.R
\name{as.sphericaldomain}
\alias{as.sphericaldomain}
\title{Convert object to a spherical domain with desired coordinate type}
\usage{
as.sphericaldomain(x, coord_type = NULL)
}
\arguments{
\item{x}{Object of class \code{"sphericaldomain"}.}

\item{coord_type}{String. Type of coordinates used for return coordinates.
Defaults to coordinate type of \code{x}. See \code{\link{spp}} for
supported coordinate types.}
}
\value{
Spherical domain of class \code{"sphericaldomain"}
}
\description{
Convert object to a spherical domain with desired coordinate type
}

