% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{as.spp}
\alias{as.spp}
\alias{as.spp.data.frame}
\alias{as.spp.default}
\alias{as.spp.matrix}
\alias{as.spp.spp}
\title{Convert Data To Class spp}
\usage{
as.spp(X, ...)

\method{as.spp}{spp}(X, ..., domain = NULL)

\method{as.spp}{data.frame}(X, ..., domain = NULL)

\method{as.spp}{matrix}(X, ..., domain = NULL)

\method{as.spp}{default}(X, ..., domain = NULL)
}
\arguments{
\item{X}{Data which will be converted into a point pattern}

\item{...}{Arguments passed to \code{\link{sphere}} if the domain is not otherwise specified.}

\item{domain}{Object of class \code{"sphericaldomain"}.}
}
\value{
An object of class \code{"spp"} (see \code{\link{spp}}) describing
  the spherical point pattern and its domain.
}
\description{
Tries to coerce any reasonable kind of data to a spherical point pattern (an
object of class \code{"spp"}) for use by the \pkg{spherespatstat} package).
}
\details{
Converts the dataset \code{X} to a point pattern (an object of class
\code{"spp"}).

This function is normally used to convert an existing point pattern dataset,
stored in another format, to the \code{"spp"} format.  To create a new point
pattern from raw data such as lat,long coordinates, it is normally easier to
use the creator function \code{\link{spp}}.

The dataset \code{X} may be:

\itemize{
  \item an object of class \code{"spp"}
  \item a matrix or data frame with at least two columns
  \item a structure with entries \code{lat}, \code{long} which are numeric
    vectors of equal length
  \item a numeric vector of length 2, interpreted as the coordinates of a
    single point.
}

The first case is typically used to change (or ensure) a specific coordinate
format by specifying either the argument \code{domain} directly or indirectly
using the argument \code{coord_type} which is passed to \code{\link{sphere}}
through the additional arguments \dots. In the last three cases the default
behaviour is to assume the domain is the entire sphere with coordinates of
type \code{"geo_deg"} (see \code{\link{spp}}). Alternatively the domain can
be specified by the argument \code{domain} or through the additional
arguments \dots.

If \code{X} is a matrix or data frame, the first and second columns will be
interpreted as the \eqn{lat} and \eqn{long} coordinates respectively. Any
additional columns will be ignored.

The function \code{as.spp} is generic, with methods for the classes
\code{"spp"}, \code{"matrix"}, \code{"data.frame"} and a default method.

Point pattern datasets can also be created by the function \code{\link{spp}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{spp}: 

\item \code{data.frame}: 

\item \code{matrix}: 

\item \code{default}: 
}}
\examples{
dom <- sphere(coord_type = "polar")
lat <- pi * c(1/4, 1/2, 3/4)
long <- 2 * pi * c(1/4, 1/2, 3/4)
X <- spp(lat, long, dom)
Y <- as.spp(X, coord_type = "geo_deg")
as.data.frame(Y)
}
\keyword{manip}
\keyword{spatial}

