% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spheredist.R
\name{bdist_spherepoints}
\alias{bdist_spherepoints}
\title{Distance to border for points inside a spherical cap}
\usage{
bdist_spherepoints(X)
}
\arguments{
\item{X}{Spherical point pattern of class \code{"spp"}}
}
\value{
Numeric
}
\description{
Distance to border for points inside a spherical cap
}

