% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.R
\name{coords.spp}
\alias{coords.spp}
\title{Extract Coordinates of a Point Pattern on a Sphere}
\usage{
\method{coords}{spp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"spp"}}

\item{...}{Ignored}
}
\value{
a \code{data.frame} with one row for each point, containing the
  coordinates.
}
\description{
Extract Coordinates of a Point Pattern on a Sphere
}

