% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.R
\name{domain.spp}
\alias{domain.spp}
\title{Extract domain of spherical point pattern}
\usage{
\method{domain}{spp}(X, ...)
}
\arguments{
\item{X}{Spherical point pattern of class \code{"spp"}.}

\item{...}{Ignored.}
}
\value{
Object of class \code{"sphericaldomain"}.
}
\description{
Extract domain of spherical point pattern
}
\examples{
lat <- c(45, 0, -45)
long <- c(90, 180, -90)
X <- spp(lat, long)
domain(X)
}

