% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sphericaldomain.R
\name{inside_sphericaldomain}
\alias{inside_sphericaldomain}
\title{Test whether points are inside spherical domain}
\usage{
inside_sphericaldomain(lat, long, domain)
}
\arguments{
\item{lat}{Latitude.}

\item{long}{Longitude.}

\item{domain}{Object of class \code{"sphericaldomain"}.}
}
\value{
Logical
}
\description{
Test whether points are inside spherical domain
}

