% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheredist.R
\name{nndist.spp}
\alias{nndist.spp}
\title{Nearest neighbour distances}
\usage{
\method{nndist}{spp}(X, ...)
}
\arguments{
\item{X}{object of class \code{"spp"}.}

\item{\dots}{ignored.}
}
\value{
Numeric of length n where n is the number of points in \code{X}.
}
\description{
Computes the vector of great circle (geodesic) distances to nearest neighbour for
the points in a point pattern on the sphere
}
\examples{
X <- spp(pi*c(1/4,1/2,3/4), 2*pi*c(1/4,1/2,3/4), domain = sphere(type = "polar"))
nndist(X)
}

