% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{npoints.spp}
\alias{npoints.spp}
\title{Extract number of points in a spherical point pattern.}
\usage{
\method{npoints}{spp}(x)
}
\arguments{
\item{x}{Spherical point patten of class \code{"spp"}.}
}
\description{
Extract number of points in a spherical point pattern.
}
\examples{
lat <- c(45, 0, -45)
long <- c(90, 180, -90)
X <- spp(lat, long)
npoints(X)
}

