% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheredist.R
\name{pairdist.spp}
\alias{pairdist.spp}
\title{Pairwise distances}
\usage{
\method{pairdist}{spp}(X, ...)
}
\arguments{
\item{X}{object of class \code{"spp"}.}

\item{\dots}{ignored.}
}
\value{
Symmetric square matrix where the (i,j)'th entry is the great circle
(geodesic) distance between points i and j.
}
\description{
Computes the matrix of great circle (geodesic) distances between all pairs
of points in a point pattern on the sphere
}
\examples{
X <- spp(pi*c(1/4,1/2,3/4), 2*pi*c(1/4,1/2,3/4), domain = sphere(type = "polar"))
pairdist(X)
}

