% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericaldomain.R
\name{print.sphericaldomain}
\alias{print.sphericaldomain}
\title{Print object of class sphericaldomain}
\usage{
\method{print}{sphericaldomain}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"sphericaldomain"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class sphericaldomain
}

