% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.R
\name{print.spp}
\alias{print.spp}
\title{Print object of class spp}
\usage{
\method{print}{spp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"spp"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class spp
}

