% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomspp.R
\name{rpoisspp}
\alias{rpoisspp}
\title{Generate Poisson point pattern on sphere}
\usage{
rpoisspp(lambda, domain = NULL, nsim = 1, ..., drop = TRUE)
}
\arguments{
\item{lambda}{Intensity. Either a single positve numeric or a vector of
length \code{nsim} indicating the intensity in each of the \code{nsim}
generated patterns.}

\item{domain}{Object of class \code{"sphericaldomain"}.}

\item{nsim}{Number of simulated realisations to be generated.}

\item{...}{Arguments passed to \code{\link{sphere}} if the domain is not otherwise specified.}

\item{drop}{Logical. If nsim=1 and drop=TRUE (the default), the result will
be a spherical point pattern, rather than a list containing a spherical
point pattern.}
}
\value{
List of \code{nsim} spherical point patterns (of class \code{"spp"}).
  Or a spherical point pattern if nsim=1 and drop=TRUE (the default).
}
\description{
Generate Poisson point pattern on sphere
}
\examples{
rpoisspp(10)

}

