% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomspp.R
\name{runifspp}
\alias{runifspp}
\title{Generate uniformly distributed point pattern on sphere}
\usage{
runifspp(n, domain = NULL, nsim = 1, ..., drop = TRUE)
}
\arguments{
\item{n}{Number of points to generate.}

\item{domain}{Object of class \code{"sphericaldomain"}.}

\item{nsim}{Number of simulated realisations to be generated.}

\item{...}{Arguments passed to \code{\link{sphere}} if the domain is not otherwise specified.}

\item{drop}{Logical. If nsim=1 and drop=TRUE (the default), the result will be a spherical point pattern, rather than a list containing a spherical point pattern.}
}
\value{
Spherical point pattern of class \code{"spp"}.
}
\description{
Generate uniformly distributed point pattern on sphere
}
\examples{
runifspp(10)
}

