% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sphericaldomain.R
\name{sphericalcap}
\alias{sphericalcap}
\title{Define spherical cap}
\usage{
sphericalcap(centre, dist, domain, ..., simplify = TRUE)
}
\arguments{
\item{centre}{Numeric of length two with latitude and longitude of point on
sphere to be used as centre for the cap.}

\item{dist}{Numeric of length one defining the extend of the cap. Should be
positive and at most half the circumfrence of the sphere.}

\item{domain}{Sphere containing the cap (object of class
\code{"sphericaldomain"}).}

\item{...}{Arguments passed to \code{\link{sphere}} to specify \code{domain}
if this is NULL (the default).}

\item{simplify}{Logical. Whether to simplify to full sphere if the cap
  extends over the full sphere.}
}
\value{
Object of class \code{"sphericaldomain"}.
}
\description{
Define spherical cap
}

