% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{spp}
\alias{spp}
\title{Spherical point pattern}
\usage{
spp(lat, long, domain = sphere(1), check = TRUE)
}
\arguments{
\item{lat}{Numeric. Latitudes of points.}

\item{long}{Numeric. Longitude of points.}

\item{domain}{Domain where the points occur. Object of class \code{"sphericaldomain"}.}

\item{check}{Logical. Check that points actually are inside the provided \code{domain}.}
}
\value{
Object of class \code{"spp"}.
}
\description{
Import spatstat
}
\examples{
dom <- sphere(coord_type = "polar")
lat <- pi * c(1/4, 1/2, 3/4)
long <- 2 * pi * c(1/4, 1/2, 3/4)
X <- spp(lat, long, dom)

dom <- sphere(coord_type = "geo_deg")
lat <- c(45, 0, -45)
long <- c(90, 180, -90)
X <- spp(lat, long, dom)
}

