% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp.R
\name{[.spp}
\alias{[.spp}
\title{Extract a subset of a point pattern on a sphere. Extraction of a subset has
the effect of thinning the points.}
\usage{
\method{[}{spp}(x, i, j, drop, ...)
}
\arguments{
\item{x}{object of class \code{"spp"}.}

\item{i}{Subset index. A valid subset index in the usual R sense, indicating
which points should be retained.}

\item{j}{Ignored. (Required for compatibility with the generic function.)}

\item{drop}{Ignored. (Required for compatibility with the generic function.)}

\item{\dots}{Ignored. (Required for compatibility with the generic
function.)}
}
\value{
object of class \code{"spp"}.
}
\description{
Extract a subset of a point pattern on a sphere. Extraction of a subset has
the effect of thinning the points.
}
\examples{
X <- spp(c(-45,0,45), c(-10,0,160))
X[1:2]
}

