% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheredist.R
\name{unitspheredist}
\alias{unitspheredist}
\title{Distance on unitsphere.}
\usage{
unitspheredist(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1, lat2}{Numeric of length 1. Latitude in radians between -pi/2 and
pi/2.}

\item{long1, long2}{Numeric of length 1. Longitude in radians between -pi and
pi (or between 0 and 2*pi).}
}
\value{
Numeric of length 1.
}
\description{
Low-level function to calculate distance on the sphere. No checking
of arguments is done.
}

